Renamed user service to farmer user service.
