+++ src/app/pages/user/login/login.component.ts
@@ -11,7 +11,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { ValidationsServiceService } from 'src/app/services/validation/validations-service.service';
+import { ValidationsServiceService } from 'src/app/services/validation-service/validations-service.service';
 // import { UserService } from 'src/app/services/users.service';
 import { ApiService } from 'src/app/_services/api-service/api.service';
 import { Subscription } from 'rxjs';
