+++ src/app/pages/user/register/register.component.ts
@@ -13,7 +13,7 @@ import { FormBuilder, FormControl, Validators } from '@angular/forms';
 import { FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
 import { ApiService } from 'src/app/_services/api-service/api.service';
-import { ValidationsServiceService } from 'src/app/services/validation/validations-service.service';
+import { ValidationsServiceService } from 'src/app/services/validation-service/validations-service.service';
 import { User } from 'src/app/_models/User';
 import { PortfolioService } from 'src/app/_services/portfolio-service/portfolio.service';
 import { Subscription } from 'rxjs';
